Move `removeImportedPoolFromUi()` into `utils\pools.ts`
- Rename `getImported` to `importedPool`
