+++ archetypes/Pools/AddAltPoolModal/index.tsx
@@ -14,7 +14,7 @@ import { BrowseTableRowData } from '../state';
 
 export default (({ open, onClose, sortedFilteredTokens }) => {
     const importPool = useStore(selectImportPool);
-    const getImported = useStore(selectImportedPools);
+    const importedPools = useStore(selectImportedPools);
     const network = useStore(selectNetwork);
 
     const [userInput, setUserInput] = useState<string>('');
@@ -34,7 +34,7 @@ export default (({ open, onClose, sortedFilteredTokens }) => {
 
     const handleImport = () => {
         const isDuplicatePool = sortedFilteredTokens.some((v: BrowseTableRowData) => v.address === userInput);
-        const isDuplicateImport = getImported.some((v) => v.address === userInput);
+        const isDuplicateImport = importedPools.some((v) => v.address === userInput);
 
         if (isDuplicatePool || isDuplicateImport) {
             setImportMsg(pool.exists);
